+++ src/helpers/__tests__/BapiClient.categories.test.ts
@@ -1,4 +1,4 @@
-import {StorefrontAPIClient} from '../../BapiClient';
+import {StorefrontAPIClient} from '../../StorefrontAPIClient';
 import {
   nockWithBapiScope,
   disableNetAndAllowBapiCors,
@@ -17,7 +17,7 @@ it.skip('Gets category by ID', async () => {
 
   const bapi = new StorefrontAPIClient({
     host: 'https://api-cloud.example.com/v1/',
-    shopId: 139,
+    countryId: 139,
   });
 
   const response = await bapi.categories.getById(20201);
@@ -41,7 +41,7 @@ it.skip('Gets categories by IDs', async () => {
 
   const bapi = new StorefrontAPIClient({
     host: 'https://api-cloud.example.com/v1/',
-    shopId: 139,
+    countryId: 139,
   });
 
   const response = await bapi.categories.getByIds([20201, 20204]);
@@ -65,7 +65,7 @@ it.skip('Gets categories by ID (including hidden)', async () => {
 
   const bapi = new StorefrontAPIClient({
     host: 'https://api-cloud.example.com/v1/',
-    shopId: 139,
+    countryId: 139,
   });
 
   const response = await bapi.categories.getById(20201, {
@@ -86,7 +86,7 @@ it.skip('Gets category by path', async () => {
 
   const bapi = new StorefrontAPIClient({
     host: 'https://api-cloud.example.com/v1/',
-    shopId: 139,
+    countryId: 139,
   });
 
   const response = await bapi.categories.getByPath(['frauen', 'bekleidung']);
@@ -105,7 +105,7 @@ it.skip('Gets root categories', async () => {
 
   const bapi = new StorefrontAPIClient({
     host: 'https://api-cloud.example.com/v1/',
-    shopId: 139,
+    countryId: 139,
   });
 
   const response = await bapi.categories.getRoots();
