+++ src/helpers/__tests__/BapiClient.countryId.test.ts
@@ -1,4 +1,4 @@
-import {StorefrontAPIClient} from '../../BapiClient';
+import {StorefrontAPIClient} from '../../StorefrontAPIClient';
 import {
   nockWithBapiScope,
   disableNetAndAllowBapiCors,
@@ -16,7 +16,7 @@ it.skip('Get basket', async () => {
 
   const bapi = new StorefrontAPIClient({
     host: 'https://api-cloud.example.com/v1/',
-    shopId: 123,
+    countryId: 123,
   });
 
   const basketKey = 'customer_2137901';
