+++ src/helpers/__tests__/BapiClient.navigation.test.ts
@@ -1,4 +1,4 @@
-import {StorefrontAPIClient} from '../../BapiClient';
+import {StorefrontAPIClient} from '../../StorefrontAPIClient';
 import {
   nockWithBapiScope,
   disableNetAndAllowBapiCors,
@@ -17,7 +17,7 @@ it.skip('Gets all navigation trees', async () => {
 
   const bapi = new StorefrontAPIClient({
     host: 'https://api-cloud.example.com/v1/',
-    shopId: 139,
+    countryId: 139,
   });
 
   const response = await bapi.navigation.getAll();
@@ -38,7 +38,7 @@ it.skip('Gets navigation tree by id', async () => {
 
   const bapi = new StorefrontAPIClient({
     host: 'https://api-cloud.example.com/v1/',
-    shopId: 139,
+    countryId: 139,
   });
 
   const response = await bapi.navigation.getById(1);
