+++ src/helpers/__tests__/BapiClient.products.test.ts
@@ -1,4 +1,4 @@
-import {StorefrontAPIClient} from '../../BapiClient';
+import {StorefrontAPIClient} from '../../StorefrontAPIClient';
 import {
   nockWithBapiScope,
   disableNetAndAllowBapiCors,
@@ -22,7 +22,7 @@ it.skip('Gets product by ID', async () => {
 
   const bapi = new StorefrontAPIClient({
     host: 'https://api-cloud.example.com/v1/',
-    shopId: 139,
+    countryId: 139,
   });
 
   const response = await bapi.products.getById(4001039);
@@ -48,7 +48,7 @@ it.skip('Gets products by IDs', async () => {
 
   const bapi = new StorefrontAPIClient({
     host: 'https://api-cloud.example.com/v1/',
-    shopId: 139,
+    countryId: 139,
   });
 
   const response = await bapi.products.getByIds([4001039, 4001029]);
@@ -68,7 +68,7 @@ it.skip('Gets products by Reference Key', async () => {
 
   const bapi = new StorefrontAPIClient({
     host: 'https://api-cloud.example.com/v1/',
-    shopId: 139,
+    countryId: 139,
   });
 
   const response = await bapi.products.getByReferenceKeys([
@@ -91,7 +91,7 @@ it.skip('Queries products', async () => {
 
   const bapi = new StorefrontAPIClient({
     host: 'https://api-cloud.example.com/v1/',
-    shopId: 139,
+    countryId: 139,
   });
 
   const response = await bapi.products.query({
@@ -118,7 +118,7 @@ it.skip('Queries products with price range', async () => {
 
   const bapi = new StorefrontAPIClient({
     host: 'https://api-cloud.example.com/v1/',
-    shopId: 139,
+    countryId: 139,
   });
 
   const response = await bapi.products.getById(3849870, {
