+++ src/helpers/__tests__/BapiClient.shopConfiguration.test.ts
@@ -1,4 +1,4 @@
-import {StorefrontAPIClient} from '../../BapiClient';
+import {StorefrontAPIClient} from '../../StorefrontAPIClient';
 import {
   nockWithBapiScope,
   disableNetAndAllowBapiCors,
@@ -30,7 +30,7 @@ it.skip('Gets product by ID', async () => {
 
   const bapi = new StorefrontAPIClient({
     host: 'https://api-cloud.example.com/v1/',
-    shopId: 139,
+    countryId: 139,
   });
 
   const response = await bapi.shopConfiguration.get();
